+++ src/main/java/net/silentchaos512/gear/api/stats/IItemStat.java
@@ -5,5 +5,23 @@ import net.minecraft.util.ResourceLocation;
 public interface IItemStat {
     ResourceLocation getStatId();
 
-    StatInstance.Operation getDefaultOperation();
+    default float getBaseValue() {
+        return 0f;
+    }
+
+    default float getDefaultValue() {
+        return 0f;
+    }
+
+    default float getMinimumValue() {
+        return Integer.MIN_VALUE;
+    }
+
+    default float getMaximumValue() {
+        return Integer.MAX_VALUE;
+    }
+
+    default StatInstance.Operation getDefaultOperation() {
+        return StatInstance.Operation.AVG;
+    }
 }