+++ src/main/java/net/silentchaos512/gear/api/stats/LazyItemStat.java
@@ -26,9 +26,4 @@ public final class LazyItemStat implements IItemStat {
     public ResourceLocation getStatId() {
         return id;
     }
-
-    @Override
-    public StatInstance.Operation getDefaultOperation() {
-        return StatInstance.Operation.AVG;
-    }
 }