+++ src/main/java/net/silentchaos512/gear/util/Const.java
@@ -114,6 +114,7 @@ public final class Const {
         public static final DataResource<ITrait> ORGANIC = DataResource.trait("organic");
         public static final DataResource<ITrait> RACKER = DataResource.trait("racker");
         public static final DataResource<ITrait> REACH = DataResource.trait("reach");
+        public static final DataResource<ITrait> RED_CARD = DataResource.trait("red_card");
         public static final DataResource<ITrait> REFRACTIVE = DataResource.trait("refractive");
         public static final DataResource<ITrait> ROAD_MAKER = DataResource.trait("road_maker");
         public static final DataResource<ITrait> RUSTIC = DataResource.trait("rustic");